requirements.txt
pdfplumber
sentence-transformers
faiss-cpu
numpy
groq
